backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
boto3==1.34.131
blinker==1.9.0
cachelib==0.13.0
certifi==2022.6.15
cffi==1.17.1
charset-normalizer==2.1.0
click==8.2.1
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
cyclopts==3.22.5
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.11.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Session==0.8.0
fonttools==4.56.0
graphviz==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.3
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mcp==1.12.3
mdurl==0.1.2
more-itertools==10.7.0
msgspec==0.19.0
numpy==2.2.3
oauthlib==3.2.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==24.2
parse==1.20.2
pathable==0.4.4
pillow==11.1.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.26.0
six==1.16.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.11
uvicorn==0.35.0
Werkzeug==3.1.1
